+++ lib/src/components/form/fields/select.dart
@@ -438,6 +438,12 @@ class _ShadFormBuilderSelectState<T>
   }
 
   @override
+  void didChange(T? value) {
+    super.didChange(value);
+    controller.value = value is T ? <T>{value} : <T>{};
+  }
+
+  @override
   void dispose() {
     controller.removeListener(onControllerChange);
     _controller?.dispose();
@@ -445,7 +451,9 @@ class _ShadFormBuilderSelectState<T>
   }
 
   void onControllerChange() {
-    didChange(controller.value.firstOrNull);
+    if (controller.value.firstOrNull != value) {
+      didChange(controller.value.firstOrNull);
+    }
   }
 
   @override